Invites added, competition fixed

+++ AuthenticationTest/Controllers/TeamsController.cs
@@ -25,6 +25,10 @@ namespace AuthenticationTest.Controllers
         // GET: Teams
         public async Task<IActionResult> Index()
         {
+            string email = User.FindFirst(ClaimTypes.Name).Value;
+            ViewBag.email = email;
+            var invites = _context.Invite.FromSqlRaw("SELECT * FROM Invite WHERE InvitedStudentEmail = ({0})", email);
+            ViewBag.invites = invites;
             return View(await _context.Team.ToListAsync());
         }
 
@@ -45,6 +49,8 @@ namespace AuthenticationTest.Controllers
 
             var members = _context.Student.FromSqlRaw("SELECT * FROM Student WHERE TeamId = ({0})", id);
             ViewBag.members = await members.ToListAsync();
+            string email = User.FindFirst(ClaimTypes.Name).Value;
+            ViewBag.email = email;
 
             return View(team);
         }
@@ -60,6 +66,7 @@ namespace AuthenticationTest.Controllers
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Student")]
         public async Task<IActionResult> Create([Bind("Id,Name")] Team team)
         {
             if (ModelState.IsValid)
@@ -67,6 +74,7 @@ namespace AuthenticationTest.Controllers
                 string _email = User.FindFirst(ClaimTypes.Name).Value;
                 // check whether the user already exists in database
                 var student = await _context.Student.FindAsync(_email);
+                team.TeamLeaderEmail = student.Email;
                 _context.Add(team);
                 student.Team = team;
                 _context.Update(student);
@@ -96,6 +104,7 @@ namespace AuthenticationTest.Controllers
 
         //JOIN TEAM POST
         [HttpPost]
+        [Authorize(Roles = "Student")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> JoinTeam(int id)
         {
@@ -107,6 +116,60 @@ namespace AuthenticationTest.Controllers
             var student = await _context.Student.FindAsync(_email);
             //Assign student to team
             student.Team = team;
+            //Save changes to DB
+            await _context.SaveChangesAsync();
+            //Redirecto to index
+            return RedirectToAction(nameof(Index));
+        }
+
+        //ACCEPT POST
+        [HttpPost]
+        [Authorize(Roles = "Student")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> RespondToInvite(int id, string submit)
+        {
+            //get invite
+            var invite = _context.Invite.Include("InvitedStudent").Include("Team").Single(x => x.Id == id);
+            switch (submit) {
+            case "Accept":
+                    //Find logged in user's email
+                    string _email = User.FindFirst(ClaimTypes.Name).Value;
+                    // get student by email
+                    var student = await _context.Student.FindAsync(_email);
+                    //Assign student to team
+                    student.Team = invite.Team;
+                    //Delete invite from DB
+                    _context.Remove(invite);
+                    //Save changes to DB
+                    await _context.SaveChangesAsync();
+                    //Redirecto to index
+                    return RedirectToAction(nameof(Index));
+            case "Decline":
+                    //Delete invite from DB
+                    _context.Remove(invite);
+                    //Save changes to DB
+                    await _context.SaveChangesAsync();
+                    //Redirecto to index
+                    return RedirectToAction(nameof(Index));
+            }
+            return RedirectToAction(nameof(Index));
+        }
+
+        //INVITE POST
+        [HttpPost]
+        [Authorize(Roles = "Student")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Invite(int id, string email)
+        {
+            //Find team by ID
+            var team = await _context.Team.FindAsync(id);
+            // get student by email
+            var student = await _context.Student.FindAsync(email);
+
+            var invite = new Invite();
+            invite.Team = team;
+            invite.InvitedStudent = student;
+            _context.Add(invite);
             //Save changes to DB
             await _context.SaveChangesAsync();
             //Redirecto to index
@@ -128,6 +191,10 @@ namespace AuthenticationTest.Controllers
             {
                 return NotFound();
             }
+            if (User.FindFirst(ClaimTypes.Name).Value != team.TeamLeaderEmail)
+            {
+                return RedirectToAction(nameof(Index));
+            }
             return View(team);
         }
 
@@ -138,6 +205,12 @@ namespace AuthenticationTest.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name")] Team team)
         {
+
+            if (User.FindFirst(ClaimTypes.Name).Value != team.TeamLeaderEmail)
+            {
+                return RedirectToAction(nameof(Index));
+            }
+
             if (id != team.Id)
             {
                 return NotFound();
@@ -180,6 +253,10 @@ namespace AuthenticationTest.Controllers
             {
                 return NotFound();
             }
+            if (User.FindFirst(ClaimTypes.Name).Value != team.TeamLeaderEmail)
+            {
+                return RedirectToAction(nameof(Index));
+            }
 
             return View(team);
         }
@@ -190,6 +267,10 @@ namespace AuthenticationTest.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var team = await _context.Team.FindAsync(id);
+            if (User.FindFirst(ClaimTypes.Name).Value != team.TeamLeaderEmail)
+            {
+                return RedirectToAction(nameof(Index));
+            }
             _context.Team.Remove(team);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
